Projectile attacks, backgrounds.
- added some preloaded sounds
- added projectile item data, projectile auto-equipping for inventory,
icon for inventory ui, projectile range for char stats, ranged attack
packet
- changed some raw pointers to optionals
- settings are mapped to an enum
- fully implemented backgrounds
- added screen fade effect for opengl mode
- added some more item movements for inventories
- added some utils

+++ IO/Components/Nametag.cpp
@@ -25,7 +25,7 @@ namespace IO
 	{
 		if (tag.count(selected))
 		{
-			int16_t width = static_cast<int16_t>(name.width());
+			int16_t width = name.width();
 			vector2d<int16_t> startpos = position - vector2d<int16_t>(8 + width / 2, -2);
 
 			using::Graphics::DrawArgument;
